Timezone/requirements.txt
discord.py>=2.0.0
pytz>=2023.3
python-dotenv>=1.0.0